+++ crates/prek/src/languages/docker_image.rs
@@ -44,7 +44,7 @@ impl LanguageImpl for DockerImage {
             .flat_map(|(key, value)| ["-e".to_owned(), format!("{key}={value}")])
             .collect();
 
-        let entry = hook.entry.resolve(None)?;
+        let entry = hook.entry.split()?;
         let run = async |batch: &[&Path]| {
             let mut cmd = Docker::docker_run_cmd(hook.work_dir());
             let mut output = cmd